Add stop_video mode.

+++ rpi_tank_rack.rb
@@ -27,6 +27,7 @@ class VideoStreamer
 
 		def stop
 			pid = self.pid
+			return if pid.nil?
 
 			5.times do
 				Process.kill(:QUIT, pid)
@@ -79,7 +80,11 @@ class WebApplication
 		(fps = @params['fps']) && fps.any? && (streaming_options[:framerate] = fps.first)
 
 		if streaming_options.any? || !VideoStreamer.running?
-			VideoStreamer.start(streaming_options)
+			if res.any? && res.first.eql?('stop')
+				VideoStreamer.stop
+			else
+				VideoStreamer.start(streaming_options)
+			end
 		end
 
 		[200, { 'Content-Type' => 'text/html' }, [ERB.new(File.read('index.html.erb')).result(binding)]]
@@ -200,6 +205,7 @@ class PowerState
 		# Autosubmit suggests we are in a programming mode; do not bother developers with re-sending commands.
 		# TODO(dotdoom): 2014-02-20: isolate this from Free Controls mode.
 		GpiodClient.send 'set_fallback_timeout 15' if @autosubmit
+		GpiodClient.send "set_fallback_timeout 15" if @autosubmit
 		GpiodClient.send "set_output #{pins}"
 	end
